+++ tooling/bundler/src/bundle/macos/sign.rs
@@ -206,7 +206,9 @@ fn try_sign(
     args.push(entitlements_path);
   }
 
-  if is_an_executable {
+  // add runtime flag by default
+
+  if is_an_executable && settings.macos().hardened_runtime {
     args.push("--options");
     args.push("runtime");
   }
@@ -305,23 +307,21 @@ pub fn notarize(
       log::info!(action = "Notarizing"; "{}", log_message);
       staple_app(app_bundle_path)?;
       Ok(())
-    } else {
-      if let Ok(output) = Command::new("xcrun")
-        .args(["notarytool", "log"])
-        .arg(&submit_output.id)
-        .notarytool_args(&auth)
-        .output_ok()
-      {
-        Err(
-          anyhow::anyhow!(
-            "{log_message}\nLog:\n{}",
-            String::from_utf8_lossy(&output.stdout)
-          )
-          .into(),
+    } else if let Ok(output) = Command::new("xcrun")
+      .args(["notarytool", "log"])
+      .arg(&submit_output.id)
+      .notarytool_args(&auth)
+      .output_ok()
+    {
+      Err(
+        anyhow::anyhow!(
+          "{log_message}\nLog:\n{}",
+          String::from_utf8_lossy(&output.stdout)
         )
-      } else {
-        Err(anyhow::anyhow!("{log_message}").into())
-      }
+        .into(),
+      )
+    } else {
+      Err(anyhow::anyhow!("{log_message}").into())
     }
   } else {
     Err(anyhow::anyhow!("failed to parse notarytool output as JSON: `{output_str}`").into())